Fix: Admin-Feedback-Seite Server-Exception bei vorhandenen Ruckmeldungen
Ursache: @neondatabase/serverless liefert numeric/ROUND(AVG(...)) als String
zuruck (Praezisionsschutz). Die als number typisierten Durchschnitte waren zur
Laufzeit Strings, StarBar rief value.toFixed(1) darauf auf -> TypeError ->
server-side exception (trat erst ab der ersten Ruckmeldung auf; bei 0 Rows sind
alle Werte null und die Seite rendert).

Fix:
- getFeedbackStats: avg_* in SQL auf ::float8 casten -> echte JS-Zahl
- StarBar: defensive Number()-Coercion + NaN-Guard
- avgGesamt-KPI: Number()-Coercion

tsc gruen.

+++ lib/db.ts
@@ -376,15 +376,18 @@ export async function upsertAssessment(
 }
 
 export async function getFeedbackStats(): Promise<FeedbackStats[]> {
+  // Wichtig: numeric/ROUND kommt über @neondatabase/serverless sonst als String
+  // zurück (Präzisionsschutz). Cast auf float8 → echte JS-Zahl (sonst bricht
+  // z. B. value.toFixed(1) im Admin-Frontend mit einer Server-Exception).
   const rows = await sql`
     SELECT
       module_id,
       COUNT(*)::int AS count,
-      ROUND(AVG(rating_inhalt)::numeric, 1) AS avg_inhalt,
-      ROUND(AVG(rating_tempo)::numeric, 1) AS avg_tempo,
-      ROUND(AVG(rating_praxis)::numeric, 1) AS avg_praxis,
-      ROUND(AVG(rating_material)::numeric, 1) AS avg_material,
-      ROUND(AVG(rating_gesamt)::numeric, 1) AS avg_gesamt
+      ROUND(AVG(rating_inhalt)::numeric, 1)::float8 AS avg_inhalt,
+      ROUND(AVG(rating_tempo)::numeric, 1)::float8 AS avg_tempo,
+      ROUND(AVG(rating_praxis)::numeric, 1)::float8 AS avg_praxis,
+      ROUND(AVG(rating_material)::numeric, 1)::float8 AS avg_material,
+      ROUND(AVG(rating_gesamt)::numeric, 1)::float8 AS avg_gesamt
     FROM feedback
     GROUP BY module_id
     ORDER BY module_id
